[fs] Reorganize the hailtop.aiotools.fs module
