+++ hail/python/hailtop/aiotools/fs/constants.py
@@ -1,0 +1,2 @@
+FILE = 'file'
+DIR = 'dir'
