+++ hail/python/hailtop/aiotools/fs/copier.py
@@ -1,446 +1,20 @@
-from typing import Any, AsyncContextManager, Optional, List, Type, BinaryIO, cast, Set, AsyncIterator, Union, Dict, Callable
-from types import TracebackType
-import abc
+from typing import TYPE_CHECKING, Any, Optional, List, Union, Dict, Callable
 import os
 import os.path
-import io
-import stat
-import shutil
 import asyncio
-from concurrent.futures import ThreadPoolExecutor
-import urllib.parse
 import functools
 import humanize
 from hailtop.utils import (
-    retry_transient_errors, blocking_to_async, url_basename, url_join, bounded_gather2,
-    time_msecs, humanize_timedelta_msecs, OnlineBoundedGather2)
-from .exceptions import FileAndDirectoryError, UnexpectedEOFError
-from .weighted_semaphore import WeightedSemaphore
-from .stream import ReadableStream, WritableStream, blocking_readable_stream_to_async, blocking_writable_stream_to_async
-
-
-class FileStatus(abc.ABC):
-    @abc.abstractmethod
-    async def size(self) -> int:
-        pass
-
-    @abc.abstractmethod
-    async def __getitem__(self, key: str) -> Any:
-        pass
-
-
-class FileListEntry(abc.ABC):
-    @abc.abstractmethod
-    def name(self) -> str:
-        pass
-
-    @abc.abstractmethod
-    async def url(self) -> str:
-        pass
-
-    @abc.abstractmethod
-    def url_maybe_trailing_slash(self) -> str:
-        pass
-
-    @abc.abstractmethod
-    async def is_file(self) -> bool:
-        pass
-
-    @abc.abstractmethod
-    async def is_dir(self) -> bool:
-        pass
-
-    @abc.abstractmethod
-    async def status(self) -> FileStatus:
-        pass
-
-
-class MultiPartCreate(abc.ABC):
-    @abc.abstractmethod
-    async def create_part(self, number: int, start: int, size_hint: Optional[int] = None) -> AsyncContextManager[WritableStream]:
-        pass
-
-    @abc.abstractmethod
-    async def __aenter__(self) -> 'MultiPartCreate':
-        pass
-
-    @abc.abstractmethod
-    async def __aexit__(self,
-                        exc_type: Optional[Type[BaseException]],
-                        exc_val: Optional[BaseException],
-                        exc_tb: Optional[TracebackType]) -> None:
-        pass
-
-
-class AsyncFS(abc.ABC):
-    FILE = 'file'
-    DIR = 'dir'
-
-    @abc.abstractmethod
-    def schemes(self) -> Set[str]:
-        pass
-
-    @abc.abstractmethod
-    async def open(self, url: str) -> ReadableStream:
-        pass
-
-    @abc.abstractmethod
-    async def open_from(self, url: str, start: int) -> ReadableStream:
-        pass
-
-    @abc.abstractmethod
-    async def create(self, url: str, *, retry_writes: bool = True) -> AsyncContextManager[WritableStream]:
-        pass
-
-    @abc.abstractmethod
-    async def multi_part_create(
-            self,
-            sema: asyncio.Semaphore,
-            url: str,
-            num_parts: int) -> MultiPartCreate:
-        pass
-
-    @abc.abstractmethod
-    async def mkdir(self, url: str) -> None:
-        pass
-
-    @abc.abstractmethod
-    async def makedirs(self, url: str, exist_ok: bool = False) -> None:
-        pass
-
-    @abc.abstractmethod
-    async def statfile(self, url: str) -> FileStatus:
-        pass
-
-    @abc.abstractmethod
-    async def listfiles(self, url: str, recursive: bool = False) -> AsyncIterator[FileListEntry]:
-        pass
-
-    @abc.abstractmethod
-    async def staturl(self, url: str) -> str:
-        pass
-
-    async def _staturl_parallel_isfile_isdir(self, url: str) -> str:
-        assert not url.endswith('/')
-
-        async def with_exception(f, *args, **kwargs):
-            try:
-                return (await f(*args, **kwargs)), None
-            except Exception as e:
-                return None, e
-
-        [(is_file, isfile_exc), (is_dir, isdir_exc)] = await asyncio.gather(
-            with_exception(self.isfile, url), with_exception(self.isdir, url + '/'))
-        # raise exception deterministically
-        if isfile_exc:
-            raise isfile_exc
-        if isdir_exc:
-            raise isdir_exc
-
-        if is_file:
-            if is_dir:
-                raise FileAndDirectoryError(url)
-            return AsyncFS.FILE
-
-        if is_dir:
-            return AsyncFS.DIR
-
-        raise FileNotFoundError(url)
-
-    @abc.abstractmethod
-    async def isfile(self, url: str) -> bool:
-        pass
-
-    @abc.abstractmethod
-    async def isdir(self, url: str) -> bool:
-        pass
-
-    @abc.abstractmethod
-    async def remove(self, url: str) -> None:
-        pass
-
-    async def _remove_doesnt_exist_ok(self, url):
-        try:
-            await self.remove(url)
-        except FileNotFoundError:
-            pass
-
-    @abc.abstractmethod
-    async def rmtree(self, sema: Optional[asyncio.Semaphore], url: str) -> None:
-        pass
-
-    async def _rmtree_with_recursive_listfiles(self, sema: asyncio.Semaphore, url: str) -> None:
-        async with OnlineBoundedGather2(sema) as pool:
-            try:
-                it = await self.listfiles(url, recursive=True)
-            except FileNotFoundError:
-                return
-            async for entry in it:
-                await pool.call(self._remove_doesnt_exist_ok, await entry.url())
-
-    async def touch(self, url: str) -> None:
-        async with await self.create(url):
-            pass
-
-    async def read(self, url: str) -> bytes:
-        async with await self.open(url) as f:
-            return await f.read()
-
-    async def read_from(self, url: str, start: int) -> bytes:
-        async with await self.open_from(url, start) as f:
-            return await f.read()
-
-    async def read_range(self, url: str, start: int, end: int) -> bytes:
-        n = (end - start) + 1
-        async with await self.open_from(url, start) as f:
-            return await f.readexactly(n)
-
-    async def write(self, url: str, data: bytes) -> None:
-        async def _write() -> None:
-            async with await self.create(url, retry_writes=False) as f:
-                await f.write(data)
-
-        await retry_transient_errors(_write)
-
-    async def exists(self, url: str) -> bool:
-        try:
-            await self.statfile(url)
-        except FileNotFoundError:
-            return False
-        else:
-            return True
-
-    async def close(self) -> None:
-        pass
-
-    async def __aenter__(self) -> 'AsyncFS':
-        return self
-
-    async def __aexit__(self,
-                        exc_type: Optional[Type[BaseException]],
-                        exc_val: Optional[BaseException],
-                        exc_tb: Optional[TracebackType]) -> None:
-        await self.close()
-
-    @staticmethod
-    def _copy_part_size():
-        '''Part size when copying using multi-part uploads.  The part size of
-        the destination filesystem is used.'''
-        return 128 * 1024 * 1024
-
-    async def copy(self,
-                   sema: asyncio.Semaphore,
-                   transfer: Union['Transfer', List['Transfer']],
-                   return_exceptions: bool = False,
-                   *,
-                   files_listener: Optional[Callable[[int], None]] = None,
-                   bytes_listener: Optional[Callable[[int], None]] = None) -> 'CopyReport':
-        copier = Copier(self)
-        copy_report = CopyReport(transfer, files_listener=files_listener, bytes_listener=bytes_listener)
-        await copier.copy(sema, copy_report, transfer, return_exceptions)
-        copy_report.mark_done()
-        return copy_report
-
-
-class LocalStatFileStatus(FileStatus):
-    def __init__(self, stat_result):
-        self._stat_result = stat_result
-        self._items = None
-
-    async def size(self) -> int:
-        return self._stat_result.st_size
-
-    async def __getitem__(self, key: str) -> Any:
-        raise KeyError(key)
-
-
-class LocalFileListEntry(FileListEntry):
-    def __init__(self, thread_pool, base_url, entry):
-        assert '/' not in entry.name
-        self._thread_pool = thread_pool
-        if not base_url.endswith('/'):
-            base_url = f'{base_url}/'
-        self._base_url = base_url
-        self._entry = entry
-        self._status = None
-
-    def name(self) -> str:
-        return self._entry.name
-
-    async def url(self) -> str:
-        trailing_slash = "/" if await self.is_dir() else ""
-        return f'{self._base_url}{self._entry.name}{trailing_slash}'
-
-    def url_maybe_trailing_slash(self) -> str:
-        return f'{self._base_url}{self._entry.name}'
-
-    async def is_file(self) -> bool:
-        return not await self.is_dir()
-
-    async def is_dir(self) -> bool:
-        return await blocking_to_async(self._thread_pool, self._entry.is_dir)
-
-    async def status(self) -> LocalStatFileStatus:
-        if self._status is None:
-            if await self.is_dir():
-                raise IsADirectoryError()
-            self._status = LocalStatFileStatus(await blocking_to_async(self._thread_pool, self._entry.stat))
-        return self._status
-
-
-class LocalMultiPartCreate(MultiPartCreate):
-    def __init__(self, fs: 'LocalAsyncFS', path: str, num_parts: int):
-        self._fs = fs
-        self._path = path
-        self._num_parts = num_parts
-
-    async def create_part(self, number: int, start: int, size_hint: Optional[int] = None):  # pylint: disable=unused-argument
-        assert 0 <= number < self._num_parts
-        f = await blocking_to_async(self._fs._thread_pool, open, self._path, 'r+b')
-        f.seek(start)
-        return blocking_writable_stream_to_async(self._fs._thread_pool, cast(BinaryIO, f))
-
-    async def __aenter__(self) -> 'LocalMultiPartCreate':
-        return self
-
-    async def __aexit__(self,
-                        exc_type: Optional[Type[BaseException]],
-                        exc_val: Optional[BaseException],
-                        exc_tb: Optional[TracebackType]) -> None:
-        if exc_val:
-            try:
-                await self._fs.remove(self._path)
-            except FileNotFoundError:
-                pass
-
-
-class LocalAsyncFS(AsyncFS):
-    def __init__(self, thread_pool: ThreadPoolExecutor, max_workers=None):
-        if not thread_pool:
-            thread_pool = ThreadPoolExecutor(max_workers=max_workers)
-        self._thread_pool = thread_pool
-
-    def schemes(self) -> Set[str]:
-        return {'file'}
-
-    @staticmethod
-    def _get_path(url):
-        parsed = urllib.parse.urlparse(url)
-        if parsed.scheme and parsed.scheme != 'file':
-            raise ValueError(f"invalid file URL: {url}, invalid scheme: expected file, got {parsed.scheme}")
-        if parsed.netloc and parsed.netloc != 'localhost':
-            raise ValueError(f"invalid file URL: {url}, invalid netloc: expected localhost or empty, got {parsed.netloc}")
-        if parsed.params:
-            raise ValueError(f"invalid file URL: params not allowed: {url}")
-        if parsed.query:
-            raise ValueError(f"invalid file URL: query not allowed: {url}")
-        if parsed.fragment:
-            raise ValueError(f"invalid file URL: fragment not allowed: {url}")
-        return parsed.path
-
-    async def open(self, url: str) -> ReadableStream:
-        f = await blocking_to_async(self._thread_pool, open, self._get_path(url), 'rb')
-        return blocking_readable_stream_to_async(self._thread_pool, cast(BinaryIO, f))
-
-    async def open_from(self, url: str, start: int) -> ReadableStream:
-        f = await blocking_to_async(self._thread_pool, open, self._get_path(url), 'rb')
-        f.seek(start, io.SEEK_SET)
-        return blocking_readable_stream_to_async(self._thread_pool, cast(BinaryIO, f))
-
-    async def create(self, url: str, *, retry_writes: bool = True) -> WritableStream:  # pylint: disable=unused-argument
-        f = await blocking_to_async(self._thread_pool, open, self._get_path(url), 'wb')
-        return blocking_writable_stream_to_async(self._thread_pool, cast(BinaryIO, f))
-
-    async def multi_part_create(
-            self,
-            sema: asyncio.Semaphore,  # pylint: disable=unused-argument
-            url: str,
-            num_parts: int) -> MultiPartCreate:
-        # create an empty file
-        # will be opened r+b to write the parts
-        async with await self.create(url):
-            pass
-        return LocalMultiPartCreate(self, self._get_path(url), num_parts)
-
-    async def statfile(self, url: str) -> LocalStatFileStatus:
-        path = self._get_path(url)
-        stat_result = await blocking_to_async(self._thread_pool, os.stat, path)
-        if stat.S_ISDIR(stat_result.st_mode):
-            raise FileNotFoundError(f'is directory: {url}')
-        return LocalStatFileStatus(stat_result)
-
-    # entries has no type hint because the return type of os.scandir
-    # appears to be a private type, posix.ScandirIterator.
-    # >>> import os
-    # >>> entries = os.scandir('.')
-    # >>> type(entries)
-    # <class 'posix.ScandirIterator'>
-    # >>> import posix
-    # >>> posix.ScandirIterator
-    # Traceback (most recent call last):
-    #   File "<stdin>", line 1, in <module>
-    # AttributeError: module 'posix' has no attribute 'ScandirIterator'
-    async def _listfiles_recursive(self, url: str, entries) -> AsyncIterator[FileListEntry]:
-        async for file in self._listfiles_flat(url, entries):
-            if await file.is_file():
-                yield file
-            else:
-                new_url = await file.url()
-                new_path = self._get_path(new_url)
-                new_entries = await blocking_to_async(self._thread_pool, os.scandir, new_path)
-                async for subfile in self._listfiles_recursive(new_url, new_entries):
-                    yield subfile
-
-    async def _listfiles_flat(self, url: str, entries) -> AsyncIterator[FileListEntry]:
-        with entries:
-            for entry in entries:
-                yield LocalFileListEntry(self._thread_pool, url, entry)
-
-    async def listfiles(self, url: str, recursive: bool = False) -> AsyncIterator[FileListEntry]:
-        path = self._get_path(url)
-        entries = await blocking_to_async(self._thread_pool, os.scandir, path)
-        if recursive:
-            return self._listfiles_recursive(url, entries)
-        return self._listfiles_flat(url, entries)
-
-    async def staturl(self, url: str) -> str:
-        path = self._get_path(url)
-        stat_result = await blocking_to_async(self._thread_pool, os.stat, path)
-        if stat.S_ISDIR(stat_result.st_mode):
-            return AsyncFS.DIR
-        return AsyncFS.FILE
-
-    async def mkdir(self, url: str) -> None:
-        path = self._get_path(url)
-        await blocking_to_async(self._thread_pool, os.mkdir, path)
-
-    async def makedirs(self, url: str, exist_ok: bool = False) -> None:
-        path = self._get_path(url)
-        await blocking_to_async(self._thread_pool, os.makedirs, path, exist_ok=exist_ok)
-
-    async def isfile(self, url: str) -> bool:
-        path = self._get_path(url)
-        return await blocking_to_async(self._thread_pool, os.path.isfile, path)
-
-    async def isdir(self, url: str) -> bool:
-        path = self._get_path(url)
-        return await blocking_to_async(self._thread_pool, os.path.isdir, path)
-
-    async def remove(self, url: str) -> None:
-        path = self._get_path(url)
-        return await blocking_to_async(self._thread_pool, os.remove, path)
-
-    async def rmtree(self, sema: Optional[asyncio.Semaphore], url: str) -> None:
-        path = self._get_path(url)
-
-        def f():
-            try:
-                shutil.rmtree(path)
-            except FileNotFoundError:
-                # no error if path does not exist
-                pass
-
-        await blocking_to_async(self._thread_pool, f)
+    retry_transient_errors, url_basename, url_join, bounded_gather2,
+    time_msecs, humanize_timedelta_msecs)
+from ..exceptions import FileAndDirectoryError, UnexpectedEOFError
+from ..weighted_semaphore import WeightedSemaphore
+
+from .constants import FILE, DIR
+
+if TYPE_CHECKING:
+    from .router_fs import RouterAsyncFS  # pylint: disable=cyclic-import
+    from .fs import MultiPartCreate, FileStatus  # pylint: disable=cyclic-import
 
 
 class Transfer:
@@ -649,7 +223,7 @@ class SourceCopier:
                          srcfile: str,
                          part_number: int,
                          this_part_size: int,
-                         part_creator: MultiPartCreate,
+                         part_creator: 'MultiPartCreate',
                          return_exceptions: bool) -> None:
         try:
             async with self.xfer_sema.acquire_manager(min(Copier.BUFFER_SIZE, this_part_size)):
@@ -675,7 +249,7 @@ class SourceCopier:
             sema: asyncio.Semaphore,
             source_report: SourceReport,
             srcfile: str,
-            srcstat: FileStatus,
+            srcstat: 'FileStatus',
             destfile: str,
             return_exceptions: bool):
         size = await srcstat.size()
@@ -714,7 +288,7 @@ class SourceCopier:
             sema: asyncio.Semaphore,
             source_report: SourceReport,
             srcfile: str,
-            srcstat: FileStatus,
+            srcstat: 'FileStatus',
             destfile: str,
             return_exceptions: bool):
         source_report.start_files(1)
@@ -739,14 +313,14 @@ class SourceCopier:
 
         if (self.treat_dest_as == Transfer.DEST_DIR
                 or (self.treat_dest_as == Transfer.INFER_DEST
-                    and dest_type == AsyncFS.DIR)):
+                    and dest_type == DIR)):
             # We know dest is a dir, but we're copying to
             # dest/basename(src), and we don't know its type.
             return url_join(self.dest, url_basename(self.src.rstrip('/'))), None
 
         if (self.treat_dest_as == Transfer.DEST_IS_TARGET
                 and self.dest.endswith('/')):
-            dest_type = AsyncFS.DIR
+            dest_type = DIR
 
         return self.dest, dest_type
 
@@ -772,10 +346,10 @@ class SourceCopier:
         if self.src_is_dir:
             raise FileAndDirectoryError(self.src)
 
-        source_report._source_type = AsyncFS.FILE
+        source_report._source_type = FILE
 
         full_dest, full_dest_type = await self._full_dest()
-        if full_dest_type == AsyncFS.DIR:
+        if full_dest_type == DIR:
             raise IsADirectoryError(full_dest)
 
         await self._copy_file_multi_part(sema, source_report, src, srcstat, full_dest, return_exceptions)
@@ -800,10 +374,10 @@ class SourceCopier:
         if self.src_is_file:
             raise FileAndDirectoryError(self.src)
 
-        source_report._source_type = AsyncFS.DIR
+        source_report._source_type = DIR
 
         full_dest, full_dest_type = await self._full_dest()
-        if full_dest_type == AsyncFS.FILE:
+        if full_dest_type == FILE:
             raise NotADirectoryError(full_dest)
 
         async def copy_source(srcentry):
@@ -875,7 +449,7 @@ class Copier:
         if (transfer.treat_dest_as == Transfer.DEST_DIR
                 or isinstance(transfer.src, list)
                 or transfer.dest.endswith('/')):
-            return AsyncFS.DIR
+            return DIR
 
         assert not transfer.dest.endswith('/')
         try:
@@ -942,99 +516,3 @@ class Copier:
                 copy_report.set_exception(e)
             else:
                 raise e
-
-
-class RouterAsyncFS(AsyncFS):
-    def __init__(self, default_scheme: Optional[str], filesystems: List[AsyncFS]):
-        scheme_fs = {}
-        schemes = set()
-        for fs in filesystems:
-            for scheme in fs.schemes():
-                if scheme not in schemes:
-                    scheme_fs[scheme] = fs
-                    schemes.add(scheme)
-
-        if default_scheme is not None and default_scheme not in schemes:
-            raise ValueError(f'default scheme {default_scheme} not in set of schemes: {", ".join(schemes)}')
-
-        self._default_scheme = default_scheme
-        self._filesystems = filesystems
-        self._schemes = schemes
-        self._scheme_fs = scheme_fs
-
-    def schemes(self) -> Set[str]:
-        return self._schemes
-
-    def _get_fs(self, url):
-        parsed = urllib.parse.urlparse(url)
-        if not parsed.scheme:
-            if self._default_scheme:
-                parsed = parsed._replace(scheme=self._default_scheme)
-            else:
-                raise ValueError(f"no default scheme and URL has no scheme: {url}")
-
-        fs = self._scheme_fs.get(parsed.scheme)
-        if fs is None:
-            raise ValueError(f"unknown scheme: {parsed.scheme}")
-
-        return fs
-
-    async def open(self, url: str) -> ReadableStream:
-        fs = self._get_fs(url)
-        return await fs.open(url)
-
-    async def open_from(self, url: str, start: int) -> ReadableStream:
-        fs = self._get_fs(url)
-        return await fs.open_from(url, start)
-
-    async def create(self, url: str, retry_writes: bool = True) -> WritableStream:
-        fs = self._get_fs(url)
-        return await fs.create(url, retry_writes=retry_writes)
-
-    async def multi_part_create(
-            self,
-            sema: asyncio.Semaphore,
-            url: str,
-            num_parts: int) -> MultiPartCreate:
-        fs = self._get_fs(url)
-        return await fs.multi_part_create(sema, url, num_parts)
-
-    async def statfile(self, url: str) -> FileStatus:
-        fs = self._get_fs(url)
-        return await fs.statfile(url)
-
-    async def listfiles(self, url: str, recursive: bool = False) -> AsyncIterator[FileListEntry]:
-        fs = self._get_fs(url)
-        return await fs.listfiles(url, recursive)
-
-    async def staturl(self, url: str) -> str:
-        fs = self._get_fs(url)
-        return await fs.staturl(url)
-
-    async def mkdir(self, url: str) -> None:
-        fs = self._get_fs(url)
-        return await fs.mkdir(url)
-
-    async def makedirs(self, url: str, exist_ok: bool = False) -> None:
-        fs = self._get_fs(url)
-        return await fs.makedirs(url, exist_ok=exist_ok)
-
-    async def isfile(self, url: str) -> bool:
-        fs = self._get_fs(url)
-        return await fs.isfile(url)
-
-    async def isdir(self, url: str) -> bool:
-        fs = self._get_fs(url)
-        return await fs.isdir(url)
-
-    async def remove(self, url: str) -> None:
-        fs = self._get_fs(url)
-        return await fs.remove(url)
-
-    async def rmtree(self, sema: Optional[asyncio.Semaphore], url: str) -> None:
-        fs = self._get_fs(url)
-        return await fs.rmtree(sema, url)
-
-    async def close(self) -> None:
-        for fs in self._filesystems:
-            await fs.close()
