+++ hail/python/hailtop/aiotools/fs/fs.py
@@ -1,0 +1,230 @@
+from typing import Any, AsyncContextManager, Optional, List, Type, Set, AsyncIterator, Union, Callable
+from types import TracebackType
+import abc
+import asyncio
+from hailtop.utils import retry_transient_errors, OnlineBoundedGather2
+from ..exceptions import FileAndDirectoryError
+from ..stream import ReadableStream, WritableStream
+from .copier import Copier, CopyReport, Transfer
+
+from .constants import FILE, DIR
+
+
+class FileStatus(abc.ABC):
+    @abc.abstractmethod
+    async def size(self) -> int:
+        pass
+
+    @abc.abstractmethod
+    async def __getitem__(self, key: str) -> Any:
+        pass
+
+
+class FileListEntry(abc.ABC):
+    @abc.abstractmethod
+    def name(self) -> str:
+        pass
+
+    @abc.abstractmethod
+    async def url(self) -> str:
+        pass
+
+    @abc.abstractmethod
+    def url_maybe_trailing_slash(self) -> str:
+        pass
+
+    @abc.abstractmethod
+    async def is_file(self) -> bool:
+        pass
+
+    @abc.abstractmethod
+    async def is_dir(self) -> bool:
+        pass
+
+    @abc.abstractmethod
+    async def status(self) -> FileStatus:
+        pass
+
+
+class MultiPartCreate(abc.ABC):
+    @abc.abstractmethod
+    async def create_part(self, number: int, start: int, size_hint: Optional[int] = None) -> AsyncContextManager[WritableStream]:
+        pass
+
+    @abc.abstractmethod
+    async def __aenter__(self) -> 'MultiPartCreate':
+        pass
+
+    @abc.abstractmethod
+    async def __aexit__(self,
+                        exc_type: Optional[Type[BaseException]],
+                        exc_val: Optional[BaseException],
+                        exc_tb: Optional[TracebackType]) -> None:
+        pass
+
+
+class AsyncFS(abc.ABC):
+    @abc.abstractmethod
+    def schemes(self) -> Set[str]:
+        pass
+
+    @abc.abstractmethod
+    async def open(self, url: str) -> ReadableStream:
+        pass
+
+    @abc.abstractmethod
+    async def open_from(self, url: str, start: int) -> ReadableStream:
+        pass
+
+    @abc.abstractmethod
+    async def create(self, url: str, *, retry_writes: bool = True) -> AsyncContextManager[WritableStream]:
+        pass
+
+    @abc.abstractmethod
+    async def multi_part_create(
+            self,
+            sema: asyncio.Semaphore,
+            url: str,
+            num_parts: int) -> MultiPartCreate:
+        pass
+
+    @abc.abstractmethod
+    async def mkdir(self, url: str) -> None:
+        pass
+
+    @abc.abstractmethod
+    async def makedirs(self, url: str, exist_ok: bool = False) -> None:
+        pass
+
+    @abc.abstractmethod
+    async def statfile(self, url: str) -> FileStatus:
+        pass
+
+    @abc.abstractmethod
+    async def listfiles(self, url: str, recursive: bool = False) -> AsyncIterator[FileListEntry]:
+        pass
+
+    @abc.abstractmethod
+    async def staturl(self, url: str) -> str:
+        pass
+
+    async def _staturl_parallel_isfile_isdir(self, url: str) -> str:
+        assert not url.endswith('/')
+
+        async def with_exception(f, *args, **kwargs):
+            try:
+                return (await f(*args, **kwargs)), None
+            except Exception as e:
+                return None, e
+
+        [(is_file, isfile_exc), (is_dir, isdir_exc)] = await asyncio.gather(
+            with_exception(self.isfile, url), with_exception(self.isdir, url + '/'))
+        # raise exception deterministically
+        if isfile_exc:
+            raise isfile_exc
+        if isdir_exc:
+            raise isdir_exc
+
+        if is_file:
+            if is_dir:
+                raise FileAndDirectoryError(url)
+            return FILE
+
+        if is_dir:
+            return DIR
+
+        raise FileNotFoundError(url)
+
+    @abc.abstractmethod
+    async def isfile(self, url: str) -> bool:
+        pass
+
+    @abc.abstractmethod
+    async def isdir(self, url: str) -> bool:
+        pass
+
+    @abc.abstractmethod
+    async def remove(self, url: str) -> None:
+        pass
+
+    async def _remove_doesnt_exist_ok(self, url):
+        try:
+            await self.remove(url)
+        except FileNotFoundError:
+            pass
+
+    @abc.abstractmethod
+    async def rmtree(self, sema: Optional[asyncio.Semaphore], url: str) -> None:
+        pass
+
+    async def _rmtree_with_recursive_listfiles(self, sema: asyncio.Semaphore, url: str) -> None:
+        async with OnlineBoundedGather2(sema) as pool:
+            try:
+                it = await self.listfiles(url, recursive=True)
+            except FileNotFoundError:
+                return
+            async for entry in it:
+                await pool.call(self._remove_doesnt_exist_ok, await entry.url())
+
+    async def touch(self, url: str) -> None:
+        async with await self.create(url):
+            pass
+
+    async def read(self, url: str) -> bytes:
+        async with await self.open(url) as f:
+            return await f.read()
+
+    async def read_from(self, url: str, start: int) -> bytes:
+        async with await self.open_from(url, start) as f:
+            return await f.read()
+
+    async def read_range(self, url: str, start: int, end: int) -> bytes:
+        n = (end - start) + 1
+        async with await self.open_from(url, start) as f:
+            return await f.readexactly(n)
+
+    async def write(self, url: str, data: bytes) -> None:
+        async def _write() -> None:
+            async with await self.create(url, retry_writes=False) as f:
+                await f.write(data)
+
+        await retry_transient_errors(_write)
+
+    async def exists(self, url: str) -> bool:
+        try:
+            await self.statfile(url)
+        except FileNotFoundError:
+            return False
+        else:
+            return True
+
+    async def close(self) -> None:
+        pass
+
+    async def __aenter__(self) -> 'AsyncFS':
+        return self
+
+    async def __aexit__(self,
+                        exc_type: Optional[Type[BaseException]],
+                        exc_val: Optional[BaseException],
+                        exc_tb: Optional[TracebackType]) -> None:
+        await self.close()
+
+    @staticmethod
+    def _copy_part_size():
+        '''Part size when copying using multi-part uploads.  The part size of
+        the destination filesystem is used.'''
+        return 128 * 1024 * 1024
+
+    async def copy(self,
+                   sema: asyncio.Semaphore,
+                   transfer: Union[Transfer, List[Transfer]],
+                   return_exceptions: bool = False,
+                   *,
+                   files_listener: Optional[Callable[[int], None]] = None,
+                   bytes_listener: Optional[Callable[[int], None]] = None) -> CopyReport:
+        copier = Copier(self)
+        copy_report = CopyReport(transfer, files_listener=files_listener, bytes_listener=bytes_listener)
+        await copier.copy(sema, copy_report, transfer, return_exceptions)
+        copy_report.mark_done()
+        return copy_report
